fix(sqlparser): semicolon in comment
