+++ src/hooks/query-code.ts
@@ -2,7 +2,7 @@ import { useCodeRunStore } from '@/store'
 import { ResultType, PromForm } from '@/store/modules/code-run/types'
 import { TableTreeParent } from '@/store/modules/database/types'
 import { EditorSelection } from '@codemirror/state'
-import { sqlFormatter } from '@/utils'
+import { sqlFormatter } from '@/utils/sql'
 import { stringType } from './types'
 
 import useSiderTabs from './sider-tabs'
